Adding your Cookie-Protecting Biz Logic
Now here’s where things get really really cool. Now in theory because of my access_control, this voter should never be called unless the user is logged in. But just in case it is let’s use is_object to check to see if the user is actually logged in.

Remember we need to do this because in Symfony 2 if you’re anonymous the user is actually set to a string. From here it’s pure business logic: if the Baker’s username equals the user’s username let’s give them access. Otherwise let’s not give them access.

So let’s refresh the 'nom' request ... and it works! We’re logged in as Ryan and we are actually nomming a Ryan cookie so this make sense.