Giving a ROLE_COOKIE_MONSTER User Special Access
Now with everything centralized I want to make things a little bit more difficult. In security.yml I’ve given the leanna user a special role called ROLE_COOKIE_MONSTER:

# app/config/security.yml
security:
    # ...

  providers:
    in_memory:
      memory:
        users:
          ryan:  { password: cookie, roles: 'ROLE_COOKIE_ENJOYER' }
          leanna: { password: cookie, roles: 'ROLE_COOKIE_MONSTER' }
If you have this role, I want to make it so you can eat any cookie even if you didn’t bake it. Seems like a jerk thing to do but let’s try it out.

To do this, we basically want to call the isGranted function on the security system from inside of our voter. Now, out-of-the-box we don’t have access to do this, so we’re going to need to do a little bit of dependency injection. If you’re thinking that we’ll inject the security.context, you’re basically right. The only issue is that because we’re inside of the security system if we try to inject the security system into here we’re going to get a circular dependency. Instead, I’m going to inject the entire container, which, yes is typically a bad practice, but in this case we can’t avoid it and it’s not going to kill us.

Head back to services.yml add an arguments key now that we have a __construct function and use @service_container to inject the entire container.

Back down in isGranted we can easily add the logic we need.

Now I’m using Symfony 2.6 which gives us a brand-new service called security.authorization_checker. This is actually a new service for Symfony 2.6. Before it was known as security.context. Now don’t worry because security.context still exists and will still exist until Symfony 3.0. So if you’re on Symfony 2.6 use the new service name. If you’re on 2.5 or earlier just use security.context. The nice thing is that both of them have the same isGranted function on it which we can use now to check to see if the user has the ROLE_COOKIE_MONSTER role. If they do, let’s give them access.

When we try it out there’s no difference and that’s a good thing. I’m logged in as ryan so I don’t actually have this role. So I’ll logout. Let’s login as leanna, password cookie, and.....COOKIES FOR EVERYBODY!